refactor: :recycle: more cluster logic out of app.js
