+++ Dockerfile
@@ -1,6 +1,6 @@
 FROM node:latest
-RUN mkdir -p /usr/src/app
-WORKDIR '/usr/src/app'
-COPY . /usr/src/app
-RUN npm ci && npm build 
+RUN mkdir -p /usr/src/main
+WORKDIR '/usr/src/main'
+COPY . /usr/src/main
+RUN npm ci && npm build
 EXPOSE 32464
